[Linux] Template deduction against eastl::fixed_string, and LP64 literals
EngineTools reaches 285 translation units with no failures.

Three headers declared their TInlineString<S> overloads as
template<size_t S> or template<eastl_size_t S>. eastl::fixed_string
declares its size parameter as int, so deducing any other type from a
TInlineString<N> fails: FileSystemPath.h, DataPath.h and
DataFileExtension.h all needed template<int S>. This is the same fix as
ResourceTypeID.h in Phase 2, and it cascaded - the "cannot initialize
object parameter" errors in the Hitbox editor were downstream of it, not
a broken inheritance chain.

DeviceRenderWorld.cpp called Math::Max( 1ULL, ...size() ). Math::Max
deduces a single type from both arguments. On Windows size_t is unsigned
long long, so they agree; Linux is LP64 where size_t is unsigned long,
and the deduction becomes ambiguous. size_t( 1 ) works on both.

TResourcePtr gains operator=( nullptr_t ), ambiguous for the same reason
as the ResourceID overload added earlier.

Two of my own bugs fixed too: the strncpy_s shim needed the MSVC template
overload that deduces the destination size from an array, and
SystemDialogs_Linux.cpp used an EE_UNUSED macro that does not exist and
was missing the TypeID header.

+++ Code/Base/FileSystem/DataPath.h
@@ -158,7 +158,10 @@ namespace EE
         inline void ReplaceExtension( String const& extension ) { ReplaceExtension( extension.c_str() ); }
 
         // Replaces the extension (excluding the '.') for this path (will create an extensions if no extension exists)
-        template<size_t S> void ReplaceExtension( TInlineString<S> const& extension ) { ReplaceExtension( extension.c_str() ); }
+        // template<int S>, not size_t or eastl_size_t: eastl::fixed_string declares its size
+        // parameter as int, so deducing any other type from a TInlineString<N> fails. MSVC
+        // accepts the narrowing during deduction. Same fix as ResourceTypeID.h.
+        template<int S> void ReplaceExtension( TInlineString<S> const& extension ) { ReplaceExtension( extension.c_str() ); }
 
         // Sub-File Paths
         //-------------------------------------------------------------------------
